We're looking at around 60,000 tsx realistically on the receiving end. Broadcasting echo brings this down to about 3000 tsx.
RequestMessage and ResolveMessage are broken! Need to implement binary marshalling-unmarshalling and fix sending the reply message

+++ pkg/carrier/mod.go
@@ -17,7 +17,7 @@ import (
 )
 
 type Carrier struct {
-	counter uint
+	counter uint64
 
 	config Config
 
@@ -244,6 +244,11 @@ func (c *Carrier) GetAddress() string {
 	return c.getCarrierToCarrierAddress()
 }
 
+// TODO
+// Rust node expects to receive uint32 to describe size of tsx
+// Implement this framing
+// Implement decoding
+// Implement reply
 func (c *Carrier) NestedPropose(P SuperBlockSummary) error {
 
 	err := c.node.GetEncoder().Encode(P)
